refactor: move ApiKeyAuthenticationSchemeHandler and extract constant for api key header name
