+++ AdminUi/src/AdminUi/AspNet/ApiKeyAuthenticationSchemeHandler.cs
@@ -3,7 +3,7 @@ using System.Text.Encodings.Web;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.Extensions.Options;
 
-namespace AdminUi.Extensions;
+namespace AdminUi.AspNet;
 
 public class ApiKeyAuthenticationSchemeOptions : AuthenticationSchemeOptions
 {
@@ -12,17 +12,19 @@ public class ApiKeyAuthenticationSchemeOptions : AuthenticationSchemeOptions
 
 public class ApiKeyAuthenticationSchemeHandler : AuthenticationHandler<ApiKeyAuthenticationSchemeOptions>
 {
+    private const string API_KEY_HEADER_NAME = "X-API-KEY";
+
     public ApiKeyAuthenticationSchemeHandler(IOptionsMonitor<ApiKeyAuthenticationSchemeOptions> options, ILoggerFactory logger, UrlEncoder encoder, ISystemClock clock) : base(options, logger, encoder, clock)
     {
     }
 
     protected override Task<AuthenticateResult> HandleAuthenticateAsync()
     {
-        var apiKey = Context.Request.Headers["X-API-KEY"];
+        var apiKey = Context.Request.Headers[API_KEY_HEADER_NAME];
 
         if (!string.IsNullOrEmpty(Options.ApiKey) && apiKey != Options.ApiKey)
         {
-            return Task.FromResult(AuthenticateResult.Fail("Invalid X-API-KEY"));
+            return Task.FromResult(AuthenticateResult.Fail($"Invalid {API_KEY_HEADER_NAME}"));
         }
         var claims = new[] { new Claim(ClaimTypes.Name, "VALID USER") };
         var identity = new ClaimsIdentity(claims, Scheme.Name);
